make it so it calculates tax.  Also change tax rate to a Double.  Also, now the package flow is working and will create a package.

+++ inventory/src/main/java/com/techvvs/inventory/model/TransactionVO.groovy
@@ -53,7 +53,7 @@ class TransactionVO implements Serializable {
     Double paid;
 
     @JsonProperty
-    Integer taxpercentage;
+    Double taxpercentage;
 
     @JsonProperty
     String notes
